backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pyjwt>=2.9.0
bcrypt>=4.2.0
python-multipart>=0.0.12
pymupdf>=1.24.0
python-docx>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.0
tiktoken>=0.7.0
aiofiles>=24.1.0
psutil>=6.0.0
pynvml>=11.5.0
transformers>=4.45.0
peft>=0.13.0
bitsandbytes>=0.44.0
accelerate>=1.0.0
torch>=2.4.0
datasets>=3.0.0
python-dotenv>=1.0.0
slowapi>=0.1.9
huggingface-hub>=0.25.0
httpx>=0.27.0
